Updated with character limits + formats.
The database tables schema has been updated with yellow boxes, which are character limits + formats for applicable fields. For example, usernames should be less than 50 characters, and the references should be in the format firstName lastName, email. This is just the present DB implementation that can be changed if we want, but I think both teams should be on the same page about these restrictions.